Add support for Quart>=0.19.0
Quart changed how it is initializing signals, removing the `AsyncNamespace` class jinja2-fragments was relying on. Hence, the jinja2-fragments was not compatible with Quart>=0.19.0, which is when this change became effective.

This commit adds support for both Quart<0.19.0 and Quart>=0.19.0.

It also changes the Quart tests to not run on Python<3.8, as Quart requires 3.8 or higher.