Remove line breaks on test html and simplify tests
The blocks of the templates have line breaks at the end of the line (as
they are not the last line of the file), hence these line breaks show up
on the rendered html for these blocks.

This was not breaking the tests because the tests using them were
removing the line breaks of the rendered html, but that feels like
tampering the output. Instead, I prefer that the data files (html)
reflect what the functions will return "as is".