Replace out-of-band tests with multiple-blocks tests
Even if htmx's out-of-band is a perfect case for using our render-blocks
functionality, using that concept on our tests and fixtures can confuse
other contributors. Additionally, the jinja templates where using
template inheritance, which adds confusion while not needed.

By generalizing these fixtures and test to just "multiple blocks", I
hope it will be easier to understand that it's just doing that.

We can keep on the documentation the mention that this capability works
great for htmx's out-of-band.